requirements.txt
torch
torchvision
transformers
safetensors
timm
numpy
pillow
opencv-python
matplotlib
tqdm
tensorboard
xgboost
